pyproject.toml
[tool.poetry]
name = "ldopa"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna"]


[tool.poetry.dependencies]
python = ">=3.8.4, <3.9"
pandas = "^1.3.4"
numpy = "^1.21.3"
plotly = "^5.3.1"

mlflow = "==1.14.1"    #"==1.20.1"   # 1.14.1
azure-storage-blob = "==12.7.1"
azure-identity ="^1.7.0"

PyYAML = "^6.0"
pyarrow = "^6.0.0"
sklearn = ">=0.0"

python-dotenv = "^0.19.1"
dash-core-components = "^2.0.0"
dash = "^2.0.0"


[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.8.0"
pre-commit = "^2.12.1"


[tool.black]
line-length = 79


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
